refactor(profile): move profile screen to its own package and remove sample files
